refactor(server): move getProjectAppName into the email utils
Drops the standalone branding module, colocating the setting reader with
getProjectSmtpConfig, the other project-level email config reader, and its tests
with the rest of the email utils tests. email/utils.ts imports nothing from auth,
so auth/utils.ts reading the app name from it adds no import cycle.
